[refactor]#25 TimeManager 추가
