+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.HomeActivity"
+            android:name=".presentation.home.HomeActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
